refactor: перенесена обработка валидаций в gateway, добавлены валидации для CommentDto, ItemRequestDto
