+++ gateway/src/main/java/ru/practicum/shareit/ShareItGateway.java
@@ -8,5 +8,4 @@ public class ShareItGateway {
 	public static void main(String[] args) {
 		SpringApplication.run(ShareItGateway.class, args);
 	}
-
 }
