+++ gateway/src/main/java/ru/practicum/shareit/booking/dto/BookItemRequestDto.java
@@ -4,6 +4,7 @@ import java.time.LocalDateTime;
 
 import javax.validation.constraints.Future;
 import javax.validation.constraints.FutureOrPresent;
+import javax.validation.constraints.NotNull;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -15,7 +16,9 @@ import lombok.NoArgsConstructor;
 public class BookItemRequestDto {
 	private long itemId;
 	@FutureOrPresent
+	@NotNull
 	private LocalDateTime start;
 	@Future
+	@NotNull
 	private LocalDateTime end;
 }
