+++ gateway/src/main/java/ru/practicum/shareit/exception/ErrorResponse.java
@@ -1,0 +1,13 @@
+package ru.practicum.shareit.exception;
+
+public class ErrorResponse {
+    private final String error;
+
+    public ErrorResponse(String error) {
+        this.error = error;
+    }
+
+    public String getError() {
+        return error;
+    }
+}
