+++ gateway/src/main/java/ru/practicum/shareit/exception/NotNullException.java
