+++ gateway/src/main/java/ru/practicum/shareit/exception/ValidationException.java
@@ -1,0 +1,7 @@
+package ru.practicum.shareit.exception;
+
+public class ValidationException extends RuntimeException {
+    public ValidationException(String message) {
+        super(message);
+    }
+}
