+++ gateway/src/main/java/ru/practicum/shareit/item/ItemController.java
@@ -2,10 +2,13 @@ package ru.practicum.shareit.item;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
+import ru.practicum.shareit.exception.NotNullException;
+import ru.practicum.shareit.exception.ValidationException;
 import ru.practicum.shareit.item.dto.CommentDto;
 import ru.practicum.shareit.item.dto.ItemDto;
 
@@ -13,6 +16,7 @@ import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Positive;
 import javax.validation.constraints.PositiveOrZero;
+import java.util.Collections;
 
 @Controller
 @RequestMapping(path = "/items")
@@ -42,6 +46,13 @@ public class ItemController {
     public ResponseEntity<Object> create(@RequestHeader("X-Sharer-User-Id") long userId,
                                           @NotNull @Valid @RequestBody ItemDto itemDto) {
         log.info("createItem method, userId {}", userId);
+        if (itemDto.getName().isBlank() || itemDto.getName() == null
+                || itemDto.getDescription() == null) {
+            throw new NotNullException("Поля имя, описание и доступность не могут быть пустыми");
+        }
+        if (itemDto.getAvailable() == null) {
+            throw new NotNullException("Поле available не может быть пустым");
+        }
         return itemClient.create(userId, itemDto);
     }
 
@@ -64,8 +75,8 @@ public class ItemController {
                                               @RequestParam(defaultValue = "0") @PositiveOrZero Integer from,
                                               @RequestParam(defaultValue = "20") @Positive Integer size) {
         log.info("createItem method");
-
-        return itemClient.search(text, from, size);
+        if (text.isBlank()) return new ResponseEntity<>(Collections.emptyList(), HttpStatus.OK);
+        else return itemClient.search(text, from, size);
     }
 
     @PostMapping("/{itemId}/comment")
@@ -73,6 +84,9 @@ public class ItemController {
                                              @Valid @RequestBody CommentDto commentDto,
                                              @RequestHeader("X-Sharer-User-Id") long authorId) {
         log.info("addComment method");
+        if (commentDto.getText().isBlank()) {
+            throw new ValidationException("Текст комментария не может быть пуст");
+        }
         return itemClient.addComment(commentDto, itemId, authorId);
     }
 }
