+++ gateway/src/main/java/ru/practicum/shareit/item/dto/CommentDto.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import javax.validation.constraints.FutureOrPresent;
+import javax.validation.constraints.NotNull;
 import java.time.LocalDateTime;
 
 @Data
@@ -13,7 +15,9 @@ public class CommentDto {
     private Integer id;
     private Integer itemId;
     private String authorName;
+    @FutureOrPresent
     private LocalDateTime created;
+    @NotNull
     private String text;
 
 }
